Fix Windows build

+++ Engine/src/graphics/renderer.cpp
@@ -1131,7 +1131,7 @@ void Renderer::updateUniformBuffer(u32 currentFrame)
     UniformBufferObject ubo{};
     ubo.model = glm::rotate(glm::mat4(1.0f), time * glm::radians(90.0f), glm::vec3(0.0f, 0.0f, 1.0f));
     ubo.view = glm::lookAt(glm::vec3(2.0f, 2.0f, 2.0f), glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f));
-    ubo.proj = glm::perspective(glm::radians(45.0f), _swapchainExtent.width / (f32)_swapchainExtent.height, 0.1f, 10.0f);
+    ubo.proj = glm::perspective(glm::radians(45.0f), (f32)_swapchainExtent.width / (f32)_swapchainExtent.height, 0.1f, 10.0f);
     // Compensate for OpenGL's Y coordinate inversion
     ubo.proj[1][1] *= -1;
